+++ source/renderer/app/containers/settings/categories/StakePoolsSettingsPage.js
@@ -20,10 +20,9 @@ export default class StakePoolsSettingsPage extends Component<InjectedProps> {
       smashServerUrlError,
       smashServerLoading,
     } = stores.staking;
+    const { isSynced, syncPercentage } = stores.networkStatus;
     const { openExternalLink } = stores.app;
     const { resetSmashServerError } = actions.staking;
-    // If `smashServerUrl` is null, waits for it to be set
-    if (!smashServerUrl) return false;
     return (
       <StakePoolsSettings
         smashServerUrl={smashServerUrl}
@@ -32,6 +31,8 @@ export default class StakePoolsSettingsPage extends Component<InjectedProps> {
         onResetSmashServerError={resetSmashServerError.trigger}
         isLoading={smashServerLoading}
         onOpenExternalLink={openExternalLink}
+        isSyncing={!isSynced}
+        syncPercentage={syncPercentage}
       />
     );
   }